backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Environment & Configuration
python-dotenv==1.0.1

# HTTP Client
httpx==0.28.1

# Data Validation
pydantic==2.8.2
pydantic-settings==2.3.4

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# Utilities
python-dateutil==2.9.0
